+++ app/admin/components/reportsDirective.js
@@ -44,6 +44,8 @@
                 vm.refreshActivity();
                 vm.refreshVisitors();
                 vm.loadApiKeys();
+                vm.filename = 'Reports_' + new Date().getTime() + '.csv';
+
             }
 
             function refreshActivity () {
@@ -65,7 +67,8 @@
             function refreshCp () {
                 commonService.getCertifiedProductActivity(vm.activityRange)
                     .then(function (data) {
-                        vm.searchedCertifiedProducts = interpretCps(data);
+                        interpretCps(data);
+                        vm.displayedCertifiedProductsUpload = [].concat(vm.searchedCertifiedProductsUpload);
                         vm.displayedCertifiedProducts = [].concat(vm.searchedCertifiedProducts);
                     });
             }
@@ -199,7 +202,8 @@
             function singleCp () {
                 commonService.getSingleCertifiedProductActivity(vm.productId)
                     .then(function (data) {
-                        vm.searchedCertifiedProducts = interpretCps(data);
+                        interpretCps(data);
+                        vm.displayedCertifiedProductsUpload = [].concat(vm.searchedCertifiedProductsUpload);
                         vm.displayedCertifiedProducts = [].concat(vm.searchedCertifiedProducts);
                     });
             }
@@ -303,7 +307,10 @@
                     {key: 'practiceType', subkey: 'name', display: 'Practice Type'},
                     {key: 'testingLab', subkey: 'name', display: 'Testing Lab'}
                 ];
-                var ret = [];
+                var output = {
+                    upload: [],
+                    other: []
+                };
                 var change;
                 var questionable;
 
@@ -320,8 +327,16 @@
                         acb: ''
                     };
                     if (data[i].description === 'Created a certified product') {
-                        activity.action = 'Created certified product <a href="#/product/' + data[i].newData.id + '">' + data[i].newData.chplProductNumber + '</a>';
+                        activity.id = data[i].newData.id;
+                        activity.chplProductNumber = data[i].newData.chplProductNumber;
                         activity.acb = data[i].newData.certifyingBody.name;
+                        activity.developer = data[i].newData.developer.name;
+                        activity.product = data[i].newData.product.name;
+                        activity.certificationEdition = data[i].newData.certificationEdition.name;
+                        activity.certificationDate = data[i].newData.certificationDate;
+                        activity.friendlyCertificationDate = new Date(activity.certificationDate).toISOString().substring(0, 10);
+                        activity.friendlyActivityDate = new Date(activity.date).toISOString().substring(0, 10);
+                        output.upload.push(activity);
                     } else if (data[i].description.startsWith('Updated certified')) {
                         questionable = data[i].activityDate > data[i].newData.certificationDate + (vm.questionableRange * 24 * 60 * 60 * 1000);
                         activity.action = 'Updated certified product <a href="#/product/' + data[i].newData.id + '">' + data[i].newData.chplProductNumber + '</a>';
@@ -366,6 +381,7 @@
                             activity.details.push('Targeted User "' + targetedUsers[j].name + '" changes<ul>' + targetedUsers[j].changes.join('') + '</ul>');
                         }
                         if (activity.details.length === 0) delete activity.details;
+                        output.other.push(activity);
                     } else if (data[i].description.startsWith('A corrective action plan for')) {
                         var cpNum = data[i].description.split(' ')[7];
                         if (data[i].description.endsWith('created.')) {
@@ -403,16 +419,19 @@
                         } else {
                             activity.action = data[i].description;
                         }
+                        output.other.push(activity);
                     } else if (data[i].description.startsWith('Documentation was added to ')) {
                         var cpNum = data[i].description.split(' ');
                         cpNum[cpNum.length - 1] = '<a href="#/product/' + data[i].newData.certifiedProductId + '">' + cpNum[cpNum.length - 1] + '</a>';
                         activity.action = cpNum.join(' ');
                         activity.acb = data[i].newData.acbName;
+                        output.other.push(activity);
                     } else if (data[i].description.startsWith('Documentation was removed from ')) {
                         var cpNum = data[i].description.split(' ');
                         cpNum[cpNum.length - 1] = '<a href="#/product/' + data[i].newData.certifiedProductId + '">' + cpNum[cpNum.length - 1] + '</a>';
                         activity.action = cpNum.join(' ');
                         activity.acb = data[i].newData.acbName;
+                        output.other.push(activity);
                     } else if (data[i].description.startsWith('Updated information for certification')) {
                         activity.action = data[i].description;
                         var capFields = [
@@ -427,12 +446,14 @@
                             change = compareItem(data[i].originalData, data[i].newData, capFields[j].key, capFields[j].display, capFields[j].filter);
                             if (change) activity.details.push(change);
                         }
+                        output.other.push(activity);
                     } else {
                         activity.action = data[i].description;
+                        output.other.push(activity);
                     }
-                    ret.push(activity);
                 }
-                return ret;
+                vm.searchedCertifiedProductsUpload = output.upload;
+                vm.searchedCertifiedProducts = output.other;
             }
 
             function compareCerts (prev, curr, questionable) {